BACKPORT: bpf: Simplify text-poke type handling
[ Upstream commit b553a6ec570044fc1ae300c6fb24f9ce204c5894 ]

Given that we have BPF_MOD_NOP_TO_{CALL,JUMP},
BPF_MOD_{CALL,JUMP}_TO_NOP
and BPF_MOD_{CALL,JUMP}_TO_{CALL,JUMP} poke types and that we also pass
in
old_addr as well as new_addr, it's a bit redundant and unnecessarily
complicates __bpf_arch_text_poke() itself since we can derive the same
from
the *_addr that were passed in. Hence simplify and use
BPF_MOD_{CALL,JUMP}
as types which also allows to clean up call-sites.

In addition to that, __bpf_arch_text_poke() currently verifies that text
matches expected old_insn before we invoke text_poke_bp(). Also add a
check
on new_insn and skip rewrite if it already matches. Reason why this is
rather
useful is that it avoids making any special casing in
prog_array_map_poke_run()
when old and new prog were NULL and has the benefit that also for this
case
we perform a check on text whether it really matches our expectations.

[Mathias Gluszczynski: Retain the generic enum and trampoline call-site
conversion. Omit the x86-only implementation and prog-array tail-call
hunks because this Exynos9810 target is arm64 and those prerequisites
are not implemented in this tree.]

Suggested-by: Andrii Nakryiko <andriin@fb.com>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Link: https://lore.kernel.org/bpf/fcb00a2b0b288d6c73de4ef58116a821c8fe8f2f.1574555798.git.daniel@iogearbox.net
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>