BACKPORT: bpf: Support bitfield reads in btf_struct_access
[ Upstream commit 976aba002fcb4b1baa71e4b0854f3d4ae48c1d4d ]

This patch allows bitfield access as a scalar.

It checks "off + size > t->size" to avoid accessing bitfield
end up accessing beyond the struct.  This check is done
outside of the loop since it is applicable to all access.

It also takes this chance to break early on the "off < moff" case.

Signed-off-by: Martin KaFai Lau <kafai@fb.com>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Acked-by: Yonghong Song <yhs@fb.com>
Link: https://lore.kernel.org/bpf/20200109003501.3855427-1-kafai@fb.com
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>