BACKPORT: bpf: Allow loading of a bpf_iter program
[ Upstream commit 15d83c4d7cef5c067a8b075ce59e97df4f60706e ]

A bpf_iter program is a tracing program with attach type
BPF_TRACE_ITER. The load attribute
  attach_btf_id
is used by the verifier against a particular kernel function,
which represents a target, e.g., __bpf_iter__bpf_map
for target bpf_map which is implemented later.

The program return value must be 0 or 1 for now.
  0 : successful, except potential seq_file buffer overflow
      which is handled by seq_file reader.
  1 : request to restart the same object

In the future, other return values may be used for filtering or
teminating the iterator.

[Mathias Gluszczynski: Keep BPF_TRACE_ITER at its upstream numeric
value because this tree omits the intervening attach types. Its newer
verifier already forces other tracing programs to return zero, so leave
iterators on [0, 1].]

Signed-off-by: Yonghong Song <yhs@fb.com>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Acked-by: Andrii Nakryiko <andriin@fb.com>
Link: https://lore.kernel.org/bpf/20200509175900.2474947-1-yhs@fb.com
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>