BACKPORT: bpf: Allow sk lookup helpers in cgroup skb
[ Upstream commit d56c2f95adb3d401bf982b6cf8fc4bb6d2f7acdd ]

Currently sk lookup helpers are allowed in tc, xdp, sk skb, and cgroup
sock_addr programs.

But they would be useful in cgroup skb as well so that for example
cgroup skb ingress program can lookup a peer socket a packet comes from
on same host and make a decision whether to allow or deny this packet
based on the properties of that socket, e.g. cgroup that peer socket
belongs to.

Allow the following sk lookup helpers in cgroup skb:
* bpf_sk_lookup_tcp;
* bpf_sk_lookup_udp;
* bpf_sk_release;

[Mathias Gluszczynski: Omit bpf_skc_lookup_tcp because this 4.9
backport does not provide that helper.]

Signed-off-by: Andrey Ignatov <rdna@fb.com>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Acked-by: Yonghong Song <yhs@fb.com>
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>