BACKPORT: flow_dissector: Pull locking up from prog attach callback
[ Upstream commit 171526f6fee84de0c39e2b7aa7e666ba0bbfd173 ]

Split out the part of attach callback that happens with attach/detach
lock acquired. This structures the prog attach callback in a way that
opens up doors for moving the locking out of flow_dissector and into
generic callbacks for attaching/detaching progs to netns in subsequent
patches.

[Mathias Gluszczynski: Apply the refactor to the Exynos9810 4.9
flow-dissector implementation and retain the already-backported global
root-namespace policy.]

Signed-off-by: Jakub Sitnicki <jakub@cloudflare.com>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Reviewed-by: Stanislav Fomichev <sdf@google.com>
Link: https://lore.kernel.org/bpf/20200531082846.2117903-2-jakub@cloudflare.com
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>