BACKPORT: flow_dissector: Move out netns_bpf prog callbacks
[ Upstream commit b27f7bb590ba835b32ef122389db158e44cfda1e ]

Move functions to manage BPF programs attached to netns that are not
specific to flow dissector to a dedicated module named
bpf/net_namespace.c.

The set of functions will grow with the addition of bpf_link support for
netns attached programs. This patch prepares ground by creating a place
for it.

This is a code move with no functional changes intended.

[Mathias Gluszczynski: Register the generic netns BPF subsystem through
Linux 4.9's pernet .exit callback instead of the unavailable .pre_exit
phase. Keep the flow-dissector-specific compatibility policy in
net/core/flow_dissector.c.]

Signed-off-by: Jakub Sitnicki <jakub@cloudflare.com>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Link: https://lore.kernel.org/bpf/20200531082846.2117903-4-jakub@cloudflare.com
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>