BACKPORT: bpf: Add socket-storage map iterator
[ Upstream commit 5ce6e77c7edf7310a0ff9532fd6b9693c082ab32 ]

The bpf iterator for bpf sock local storage map
is implemented. User space interacts with sock
local storage map with fd as a key and storage value.
In kernel, passing fd to the bpf program does not
really make sense. In this case, the sock itself is
passed to bpf program.

Signed-off-by: Yonghong Song <yhs@fb.com>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Link: https://lore.kernel.org/bpf/20200723184116.590602-1-yhs@fb.com
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>