BACKPORT: tcp: Add saw_unknown to struct tcp_options_received
[ Upstream commit 7656d68455891f7fc6689f95415fd59e7a1d629b ]

In a later patch, the bpf prog only wants to be called to handle
a header option if that particular header option cannot be handled by
the kernel.  This unknown option could be written by the peer's bpf-prog.
It could also be a new standard option that the running kernel does not
support it while a bpf-prog can handle it.

This patch adds a "saw_unknown" bit to "struct tcp_options_received"
and it uses an existing one byte hole to do that.  "saw_unknown" will
be set in tcp_parse_options() if it sees an option that the kernel
cannot handle.

[Mathias Gluszczynski: Used the existing padding byte in the Linux 4.9
TCP option state and marked unsupported experimental options without
the later SMC parser.]

Signed-off-by: Martin KaFai Lau <kafai@fb.com>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Reviewed-by: Eric Dumazet <edumazet@google.com>
Acked-by: John Fastabend <john.fastabend@gmail.com>
Link: https://lore.kernel.org/bpf/20200820190033.2884430-1-kafai@fb.com

Signed-off-by: Mathias Gluszczynski <admin@krazey.de>