BACKPORT: net: Allow iterating sockmap and sockhash
[ Upstream commit 0365351524d7560d8ed7a42801a15252c6c56f41 ]

Add bpf_iter support for sockmap / sockhash, based on the bpf_sk_storage
and
hashtable implementation. sockmap and sockhash share the same iteration
context: a pointer to an arbitrary key and a pointer to a socket. Both
pointers may be NULL, and so BPF has to perform a NULL check before
accessing
them. Technically it's not possible for sockhash iteration to yield a
NULL
socket, but we ignore this to be able to use a single iteration point.

Iteration will visit all keys that remain unmodified during the lifetime
of
the iterator. It may or may not visit newly added ones.

Switch from using rcu_dereference_raw to plain rcu_dereference, so we
gain
another guard rail if CONFIG_PROVE_RCU is enabled.

[Mathias Gluszczynski: Adapt the hash walker to this kernel's bpf_htab
types and retain the older sockmap layout.]

Signed-off-by: Lorenz Bauer <lmb@cloudflare.com>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Acked-by: Yonghong Song <yhs@fb.com>
Link: https://lore.kernel.org/bpf/20200909162712.221874-3-lmb@cloudflare.com
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>