BACKPORT: tcp: reflect tos value received in SYN to the socket
[ Upstream commit ac8f1710c12bb4c3626280ce03f05459ba8feef6 ]

This commit adds a new TCP feature to reflect the tos value received in
SYN, and send it out on the SYN-ACK, and eventually set the tos value of
the established socket with this reflected tos value. This provides a
way to set the traffic class/QoS level for all traffic in the same
connection to be the same as the incoming SYN request. It could be
useful in data centers to provide equivalent QoS according to the
incoming request.
This feature is guarded by /proc/sys/net/ipv4/tcp_reflect_tos, and is by
default turned off.

[Mathias Gluszczynski: Adapted TOS reflection to the Exynos9810 4.9
per-network sysctl table and TCP child paths.]

Signed-off-by: Wei Wang <weiwan@google.com>
Signed-off-by: Eric Dumazet <edumazet@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

Signed-off-by: Mathias Gluszczynski <admin@krazey.de>