BACKPORT: bpf: Mutex protect used_maps array and count
[ Upstream commit 984fe94f94756dacb3c8cc52904a23adf9e04da1 ]

To support modifying the used_maps array, we use a mutex to protect
the use of the counter and the array. The mutex is initialized right
after the prog aux is allocated, and destroyed right before prog
aux is freed. This way we guarantee it's initialized for both cBPF
and eBPF.

[Mathias Gluszczynski: Add the mutex to this tree's older program aux
layout and protect its existing tail-call check, instruction-dump map
lookup, and program-info map ID readers. The upstream NFP and generic
XDP readers are not present in this source state.]

Signed-off-by: YiFei Zhu <zhuyifei@google.com>
Signed-off-by: Stanislav Fomichev <sdf@google.com>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Acked-by: Andrii Nakryiko <andriin@fb.com>
Cc: YiFei Zhu <zhuyifei1999@gmail.com>
Link: https://lore.kernel.org/bpf/20200915234543.3220146-2-sdf@google.com
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>