BACKPORT: bpf: Make BTF pointer type checking generic
[ Upstream commit d7b9454a4f6333bf145189b8e769011d15bdd50e ]

Perform BTF type checks if the register we're working on contains a BTF
pointer, rather than if the argument is for a BTF pointer. This is
easier
to understand, and allows removing the code from the arg_type checking
section of the function.

[Mathias Gluszczynski: Keep the Exynos9810 4.9 verifier's runtime
helper BTF-ID resolver. The later per-argument .BTF_ids conversion will
replace that interface in its own backport.]

Signed-off-by: Lorenz Bauer <lmb@cloudflare.com>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Acked-by: Martin KaFai Lau <kafai@fb.com>
Link: https://lore.kernel.org/bpf/20200921121227.255763-6-lmb@cloudflare.com
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>