BACKPORT: bpf: Set helper raw mode close to its use
[ Upstream commit a2bbe7cc90755283f1db719eb757616cefd2a9fd ]

If we encounter a pointer to memory, we set meta->raw_mode depending
on the type of memory we point at. What isn't obvious is that this
information is only used when the next memory size argument is
encountered.

Move the assignment closer to where it's used, and add a comment that
explains what is going on.

[Mathias Gluszczynski: Apply the refactor to the Exynos9810 4.9
verifier's existing helper argument flow.]

Signed-off-by: Lorenz Bauer <lmb@cloudflare.com>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Acked-by: Martin KaFai Lau <kafai@fb.com>
Link: https://lore.kernel.org/bpf/20200921121227.255763-9-lmb@cloudflare.com
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>