BACKPORT: bpf: Use a table for helper argument checks
[ Upstream commit f79e7ea571732a6e16f15c6e2f000c347e2d7431 ]

The mapping between bpf_arg_type and bpf_reg_type is encoded in a big
hairy if statement that is hard to follow. The debug output also leaves
to be desired: if a reg_type doesn't match we only print one of the
options, instead printing all the valid ones.

Convert the if statement into a table which is then used to drive type
checking. If none of the reg_types match we print all options, e.g.:

    R2 type=rdonly_buf expected=fp, pkt, pkt_meta, map_value

[Mathias Gluszczynski: Extend the tables for the Exynos9810 4.9 tree's
map-key, allocated-memory, callback-function, and nullable stack
arguments. Retain the existing runtime BTF-ID validation after the
common register compatibility check.]

Signed-off-by: Lorenz Bauer <lmb@cloudflare.com>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Acked-by: Martin KaFai Lau <kafai@fb.com>
Link: https://lore.kernel.org/bpf/20200921121227.255763-12-lmb@cloudflare.com
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>