BACKPORT: bpf: Deref map in BPF_PROG_BIND_MAP when it's already used
[ Upstream commit 1028ae4069991e26d1522e957939fb61d2da1d12 ]

We are missing a deref for the case when we are doing BPF_PROG_BIND_MAP
on a map that's being already held by the program.
There is 'if (ret) bpf_map_put(map)' below which doesn't trigger
because we don't consider this an error.
Let's add missing bpf_map_put() for this specific condition.

[Mathias Gluszczynski: Apply the reference fix to the 4.9 fd-based map
acquisition path introduced by the preceding backport.]

Fixes: ef15314aa5de ("bpf: Add BPF_PROG_BIND_MAP syscall")
Reported-by: Alexei Starovoitov <ast@kernel.org>
Signed-off-by: Stanislav Fomichev <sdf@google.com>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Acked-by: Andrii Nakryiko <andriin@fb.com>
Link: https://lore.kernel.org/bpf/20201003002544.3601440-1-sdf@google.com
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>