BACKPORT: bpf, sockmap: Allow skipping sk_skb parser program
[ Upstream commit ef5659280eb13e8ac31c296f58cfdfa1684ac06b ]

Currently, we often run with a nop parser namely one that just does
this, 'return skb->len'. This happens when either our verdict program
can handle streaming data or it is only looking at socket data such
as IP addresses and other metadata associated with the flow. The second
case is common for a L3/L4 proxy for instance.

So lets allow loading programs without the parser then we can skip
the stream parser logic and avoid having to add a BPF program that
is effectively a nop.

[Mathias Gluszczynski: Move the common verdict path outside Linux
4.9's optional stream-parser block and adapt map linking to the
existing psock program fields and TCP protocol reinitialization.]

Signed-off-by: John Fastabend <john.fastabend@gmail.com>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Link: https://lore.kernel.org/bpf/160239297866.8495.13345662302749219672.stgit@john-Precision-5820-Tower
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>