BACKPORT: bpf: Retrieve SOL_SOCKET options from sock_addr
[ Upstream commit bcd6f4a8bedabac6949d05e418e73a0a1b309e84 ]

The _bpf_setsockopt() is able to set some of the SOL_SOCKET level
options,
however, _bpf_getsockopt() has little support to actually retrieve them.
This small patch adds few misc options such as SO_MARK, SO_PRIORITY and
SO_BINDTOIFINDEX. For the latter getter and setter are added. The mark
and
priority in particular allow to retrieve the options from BPF cgroup
hooks
to then implement custom behavior / settings on the syscall hooks
compared
to other sockets that stick to the defaults, for example.

[Mathias Gluszczynski: SO_BINDTOIFINDEX does not exist in this 4.9
socket ABI and is omitted.]

Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Acked-by: Yonghong Song <yhs@fb.com>
Link: https://lore.kernel.org/bpf/cba44439b801e5ddc1170e5be787f4dc93a2d7f9.1610406333.git.daniel@iogearbox.net
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>