BACKPORT: bpf: Add bitwise atomic instructions
[ Upstream commit 981f94c3e92146705baf97fb417a5ed1ab1a79a5 ]

This adds instructions for

atomic[64]_[fetch_]and
atomic[64]_[fetch_]or
atomic[64]_[fetch_]xor

All these operations are isomorphic enough to implement with the same
verifier, interpreter, and x86 JIT code, hence being a single commit.

The main interesting thing here is that x86 doesn't directly support
the fetch_ version these operations, so we need to generate a CMPXCHG
loop in the JIT. This requires the use of two temporary registers,
IIUC it's safe to use BPF_REG_AX and x86's AUX_REG for this purpose.

[Mathias Gluszczynski: Adapted the x86 CMPXCHG loop to the Exynos9810
4.9 JIT by adding local move and load emitters. Included the final
upstream merge correction that recognizes BPF_AND in disassembly.]

Signed-off-by: Brendan Jackman <jackmanb@google.com>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Acked-by: Yonghong Song <yhs@fb.com>
Link: https://lore.kernel.org/bpf/20210114181751.768687-10-jackmanb@google.com
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>