BACKPORT: bpf, x86: Fix BPF_FETCH atomic and/or/xor with r0 as src
[ Upstream commit b29dd96b905f3dd543f4ca729447286adf934dd6 ]

This code generates a CMPXCHG loop in order to implement atomic_fetch
bitwise operations. Because CMPXCHG is hard-coded to use rax (which
holds the BPF r0 value), it saves the _real_ r0 value into the
internal "ax" temporary register and restores it once the loop is
complete.

In the middle of the loop, the actual bitwise operation is performed
using src_reg. The bug occurs when src_reg is r0: as described above,
r0 has been clobbered and the real r0 value is in the ax register.

Therefore, perform this operation on the ax register instead, when
src_reg is r0.

[Mathias Gluszczynski: Applied the x86 JIT fix to the Exynos9810 4.9
layout. The newer split verifier-test files are absent from this tree.]

Fixes: 981f94c3e921 ("bpf: Add bitwise atomic instructions")
Signed-off-by: Brendan Jackman <jackmanb@google.com>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>
Acked-by: KP Singh <kpsingh@kernel.org>
Link: https://lore.kernel.org/bpf/20210216125307.1406237-1-jackmanb@google.com
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>