BACKPORT: skmsg: Move sk_redir from TCP_SKB_CB to skb
[ Upstream commit e3526bb92a2084cdaec6cb2855bcec98b280426c ]

Currently TCP_SKB_CB() is hard-coded in skmsg code, it certainly
does not work for any other non-TCP protocols. We can move them to
skb ext, but it introduces a memory allocation on fast path.

Fortunately, we only need to a word-size to store all the information,
because the flags actually only contains 1 bit so can be just packed
into the lowest bit of the "pointer", which is stored as unsigned
long.

Inside struct sk_buff, '_skb_refdst' can be reused because skb dst is
no longer needed after ->sk_data_ready() so we can just drop it.

[Mathias Gluszczynski: Overlay Linux 4.9's _skb_refdst with the packed
redirect state and retain a second low bit for its existing strparser
marker. Drop the destination before reusing the storage and clear it on
queued-SKB teardown.]

Signed-off-by: Cong Wang <cong.wang@bytedance.com>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Acked-by: John Fastabend <john.fastabend@gmail.com>
Acked-by: Jakub Sitnicki <jakub@cloudflare.com>
Link: https://lore.kernel.org/bpf/20210223184934.6054-5-xiyou.wangcong@gmail.com
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>