BACKPORT: bpf: Change return value of verifier function add_subprog()
[ Upstream commit 282a0f46d6cda7cf843cd77c9b53b4d1d9e31302 ]

Currently, verifier function add_subprog() returns 0 for success
and negative value for failure. Change the return value
to be the subprog number for success. This functionality will be
used in the next patch to save a call to find_subprog().

[Mathias Gluszczynski: This tree sorts the subprogram table after every
insertion, so resolve the index after sorting. Also use the upstream
capacity check before adding the synthetic exit entry.]

Signed-off-by: Yonghong Song <yhs@fb.com>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Acked-by: Andrii Nakryiko <andrii@kernel.org>
Link: https://lore.kernel.org/bpf/20210226204924.3884848-1-yhs@fb.com
Signed-off-by: Mathias Gluszczynski <admin@krazey.de>